feat: adding bottom bar in home screen - also changing folder structure of home screen
