+++ lib/presentation/home/home_screen/controller/home_controller.dart
@@ -1,0 +1,5 @@
+import 'package:get/get.dart';
+
+class HomeController extends GetxController {
+  RxInt bottomBarIndex = 0.obs;
+}
