--- lib/presentation/home_screen/binding/home_binding.dart
@@ -1,10 +1,0 @@
-import 'package:get/get.dart';
-
-import '../controller/home_controller.dart';
-
-class HomeBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut(() => HomeController());
-  }
-}
