--- lib/presentation/home_screen/controller/home_controller.dart
@@ -1,3 +1,0 @@
-import 'package:get/get.dart';
-
-class HomeController extends GetxController {}
