--- lib/presentation/home_screen/home_screen.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-
-import 'controller/home_controller.dart';
-
-class HomeScreen extends StatefulWidget {
-  const HomeScreen({super.key});
-
-  @override
-  State<HomeScreen> createState() => _HomeScreenState();
-}
-
-class _HomeScreenState extends State<HomeScreen> {
-  late final HomeController _homeController;
-
-  @override
-  void initState() {
-    _homeController = Get.find();
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: SafeArea(
-        child: Center(
-          child: Text('Home screen'),
-        ),
-      ),
-    );
-  }
-}
